[Refactor] 패키지 구조 변경
`+` domain: 도메인 모델 및 서비스
`+` infra: 도메인 서비스의 구현
