+++ src/main/java/codesquad/infra/InMemoryArticleDao.java
@@ -1,4 +1,7 @@
-package codesquad.article;
+package codesquad.infra;
+
+import codesquad.domain.article.Article;
+import codesquad.domain.article.ArticleDao;
 
 import java.util.List;
 import java.util.Optional;
