+++ src/main/java/codesquad/servlet/UsersServlet.java
@@ -1,8 +1,8 @@
 package codesquad.servlet;
 
 import codesquad.exception.DuplicateIdException;
-import codesquad.user.User;
-import codesquad.user.UserDao;
+import codesquad.domain.user.User;
+import codesquad.domain.user.UserDao;
 import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletException;
