+++ src/main/webapp/WEB-INF/views/user/list.jsp
@@ -1,6 +1,6 @@
 <%@ page contentType="text/html; charset=UTF-8" %>
 <%@ page import="java.util.List" %>
-<%@ page import="codesquad.user.User" %>
+<%@ page import="codesquad.domain.user.User" %>
 <!DOCTYPE html>
 <html lang="kr">
 <head>
